requirements.txt
pyttsx3==2.90
speechrecognition==3.10.0
openai-whisper==20231117
requests==2.32.3
torch==2.4.1
urllib3==2.2.3
tk==0.1.0
aiohttp==3.10.5